Remove more duplicate test code
