+++ src/libraries/System.IO.FileSystem/tests/System.IO.FileSystem.Tests.csproj
@@ -17,7 +17,6 @@
     <Compile Include="Base\SymbolicLinks\BaseSymbolicLinks.cs" />
     <Compile Include="Base\SymbolicLinks\BaseSymbolicLinks.FileSystem.cs" />
     <Compile Include="Base\SymbolicLinks\BaseSymbolicLinks.FileSystemInfo.cs" />
-    <Compile Include="Base\SymbolicLinks\BaseSymbolicLinks.FilesAndDirectories.cs" />
     <Compile Include="Directory\EnumerableTests.cs" />
     <Compile Include="Directory\SymbolicLinks.cs" />
     <Compile Include="DirectoryInfo\SymbolicLinks.cs" />
